refactor : Todo and Task related components
