--- src/ui/components/common/card/CardDeleteModal.tsx
@@ -1,7 +1,0 @@
-import React from "react"
-
-function CardDeleteModal() {
-  return <div>CardDeleteModal</div>
-}
-
-export default CardDeleteModal
